move artifact_registry to it's own file
